backend/pyproject.toml
[project]
name = "semantic-stocks-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "httpx>=0.28.0",
    "massive>=2.0.2",
    "pydantic>=2.10.0",
    "perplexityai>=0.22.2",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "asyncpg>=0.30.0",
    "rich>=13.9.0",
    "pandas>=2.3.3",
    "lxml>=5.0.0",  # Required for pandas.read_html() to parse Wikipedia tables
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "ruff>=0.14.9",
]

[tool.mypy]
plugins = ['pydantic.mypy']
files = ["app"]
python_version = "3.13"
explicit_package_bases = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

# Consolidated overrides for libraries without type stubs
[[tool.mypy.overrides]]
module = ["massive", "massive.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["asyncpg", "asyncpg.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["perplexity", "perplexity.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pandas", "pandas.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
    # flake8-async
    "ASYNC",
    # flake8-use-pathlib
    "PTH",
    # flake8-return
    "RET",
    # perflint
    "PERF",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Avoid specifying long messages outside the exception class
    "TRY003",
    # Consider moving statement to else block
    "TRY300",
]

[tool.ruff.format]
preview = true
